+++ core-ng/src/main/java/core/framework/api/module/SiteConfig.java
@@ -1,10 +1,17 @@
 package core.framework.api.module;
 
 import core.framework.api.http.HTTPMethod;
+import core.framework.api.util.Exceptions;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.web.ControllerHolder;
+import core.framework.impl.web.site.StaticDirectoryController;
+import core.framework.impl.web.site.StaticFileController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
 
 /**
  * @author neo
@@ -30,9 +37,18 @@ public final class SiteConfig {
         context.httpServer.siteManager.templateManager.add(path, modelClass);
     }
 
-    public void staticContent(String root) {
-        logger.info("add static content root, root={}", root);
-        context.httpServer.handler.route.add(HTTPMethod.GET, root + "/:path(*)", new ControllerHolder(context.httpServer.siteManager.staticContentController(root), true));
+    public void staticContent(String path) {
+        logger.info("add static content path, path={}", path);
+        Path contentPath = context.httpServer.siteManager.webDirectory.path(path);
+
+        if (!Files.exists(contentPath, LinkOption.NOFOLLOW_LINKS))
+            throw Exceptions.error("path does not exist, path={}", path);
+
+        if (Files.isDirectory(contentPath)) {
+            context.httpServer.handler.route.add(HTTPMethod.GET, path + "/:path(*)", new ControllerHolder(new StaticDirectoryController(contentPath), true));
+        } else {
+            context.httpServer.handler.route.add(HTTPMethod.GET, path, new ControllerHolder(new StaticFileController(contentPath), true));
+        }
     }
 
     public void cdnHosts(String... hosts) {