+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 3.6.8 (10/14/2014)
 * gradle: checkstyle update to 6.11.2, finally support intention for lambda
+* web: static content supports file under root directly (like robot.txt and favicon.ico)
+* web: static content removed 304 support, because static content is always handled by nginx in server env.
 
 ### 3.6.7 (10/12/2015 - 10/13/2015)
 * web: update undertow to 1.3.0.Final